backend-env/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
langchain
openai
faiss-cpu
python-dotenv
